end of work

+++ src/modules/personal/services/personal.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PersonalEntity } from '../personal.entity';
 import { Repository } from 'typeorm';
@@ -12,6 +12,8 @@ import { SubunitEntity } from 'src/modules/subunit/subunit.entity';
 import { CreatePersonalDto } from '../dtos/create-personal.dto';
 import { UpdatePersonalDto } from '../dtos/update-personal.dto';
 import * as moment from 'moment';
+import { SubunitService } from 'src/modules/subunit/services/subunit.service';
+import { PhysicalFaceService } from 'src/modules/physical_face/services/physical_face.service';
 
 @Injectable()
 export class PersonalService {
@@ -24,6 +26,8 @@ export class PersonalService {
     private readonly physicalFaceRepository: Repository<PhysicalFaceEntity>,
     @InjectRepository(SubunitEntity)
     private readonly subunitRepository: Repository<SubunitEntity>,
+    @Inject(PhysicalFaceService)
+    private readonly physicalFaceService: PhysicalFaceService
   ) {}
 
   async getAll() {
@@ -58,7 +62,20 @@ export class PersonalService {
 
     const newPerson: IPersonalEntity = new PersonalEntity();
 
+
+
     try {
+
+
+      if(physical_face_inn &&
+        !this.physicalFaceService.isValidInn(physical_face_inn)
+        ){
+        throw new HttpException(
+          'Your inn must be 10 numbers', HttpStatus.BAD_REQUEST
+        )
+      }
+      
+
       const physicalFace: IPhysicalFaceEntity =
         await this.physicalFaceRepository.findOne({
           where: { inn: physical_face_inn },
@@ -77,16 +94,20 @@ export class PersonalService {
           newPerson.physical_face = physicalFace;
         }
       } else {
-        throw new Error(
+        throw new HttpException(
           `Physical_face with inn: ${physical_face_inn}, doesn\'t exists. Create it or write another one physical_face_inn`,
+          HttpStatus.NOT_FOUND
         );
       }
 
       if (subunit) {
+
         newPerson.subunit = subunit;
+        
       } else {
-        throw new Error(
+        throw new HttpException(
           `Subunit_name with name: ${subunit_name}, doesn\'t exists. Create it or write another one subunit_name`,
+          HttpStatus.NOT_FOUND
         );
       }
 
@@ -100,7 +121,7 @@ export class PersonalService {
         const errorMessage = `No such position with name ${position_name}. You could chose from:[${positions.map(
           ({ name }) => `${name}`,
         )}]`;
-        throw new Error(errorMessage);
+        throw new HttpException(errorMessage, HttpStatus.NOT_FOUND);
       }
 
       if (salary) {
@@ -109,9 +130,12 @@ export class PersonalService {
 
       if (begin_date) {
         if (!moment(begin_date, 'DD.MM.YYYY', true).isValid()) {
-          throw new Error(
+          throw new HttpException(
             `This begin_date is incorrect pls use this pattern: 'DD.MM.YYYY' `,
+            HttpStatus.BAD_REQUEST
           );
+        } else {
+          newPerson.begin_date = begin_date;
         }
       }
 
@@ -122,16 +146,36 @@ export class PersonalService {
   }
 
   async updateOne(id: number, body: UpdatePersonalDto) {
-    try {
-      const {
-        physical_face_inn,
-        position_name,
-        subunit_name,
-        salary,
-        begin_date,
-      } = body;
+
+    const {
+      physical_face_inn,
+      position_name,
+      subunit_name,
+      salary,
+      begin_date,
+    } = body;
+
+    try {      
 
       const updatedPerson = new PersonalEntity();
+
+      if(
+        begin_date &&
+        !this.physicalFaceService.checkAccessOnWork(begin_date.toLocaleString().split(',')[0])
+        ){
+          throw new HttpException(
+            'You must have 18+ age for get a job', HttpStatus.BAD_REQUEST
+          )
+        }
+
+      if(
+        physical_face_inn &&
+        !this.physicalFaceService.isValidInn(physical_face_inn)
+        ){
+        throw new HttpException(
+          'Your inn must be 10 numbers', HttpStatus.BAD_REQUEST
+        )
+      }
 
       const physicalFace: IPhysicalFaceEntity =
         await this.physicalFaceRepository
@@ -149,16 +193,19 @@ export class PersonalService {
           updatedPerson.physical_face = physicalFace;
         }
       } else {
-        throw new Error(
+        throw new HttpException(
           `Physical_face with inn: ${physical_face_inn}, doesn\'t exists. Create it or write another one physical_face_inn`,
+          HttpStatus.NOT_FOUND
         );
       }
 
       if (subunit) {
         updatedPerson.subunit = subunit;
+
       } else {
-        throw new Error(
+        throw new HttpException(
           `Subunit_name with name: ${subunit_name}, doesn\'t exists. Create it or write another one subunit_name`,
+          HttpStatus.NOT_FOUND
         );
       }
 
@@ -172,7 +219,7 @@ export class PersonalService {
         const errorMessage = `No such position with name ${position_name}. You could chose from:[${positions.map(
           ({ name }) => `${name}`,
         )}]`;
-        throw new Error(errorMessage);
+        throw new HttpException(errorMessage, HttpStatus.CONFLICT);
       }
 
       if (salary) {
@@ -181,17 +228,19 @@ export class PersonalService {
 
       if (begin_date) {
         if (!moment(begin_date, 'DD.MM.YYYY', true).isValid()) {
-          throw new Error(
+          throw new HttpException(
             `This begin_date is incorrect pls use this pattern: 'DD.MM.YYYY' `,
+            HttpStatus.BAD_REQUEST
           );
+        } else {
+          updatedPerson.begin_date = begin_date
         }
       }
 
       return await this.personalRepository.update(
         { id },
         {
-          ...updatedPerson,
-          begin_date,
+          ...updatedPerson
         },
       );
     } catch (err) {
@@ -201,11 +250,14 @@ export class PersonalService {
 
   async deleteOne(id: number) {
     try {
-      // const currentSubunit = await this.personalRepository.findOne({where: {id}})
-      // .then(({id}) => this.subunitRepository.findOne({where: {id}}))
-      // if(currentSubunit){
-      //   throw new Error(`Set to null supervisor_id in subunit table by id: ${currentSubunit.id}, for fire this person. Because he is supervisor`)
-      // }
+      const currentSubunit = await this.personalRepository.findOne({where: {id}})
+      .then(({id}) => this.subunitRepository.findOne({where: {id}}))
+      if(currentSubunit){
+        throw new HttpException(
+          `Set to null supervisor_id in subunit table by id: ${currentSubunit.id}, for fire this person. Because he is supervisor`,
+          HttpStatus.CONFLICT
+        )
+      }
       return await this.personalRepository.delete({ id });
     } catch (err) {
       throw err;
